perf: runtime optimizations across COREMOD_arith and COREMOD_memory (#91)
Apply several targeted performance improvements guided by
performance-practices.md:

1. __builtin_memcpy in stream processors
   Replace libc memcpy with __builtin_memcpy in hot stream-copy paths
   (stream_UDP.c, stream_TCP.c, stream_delay.c, im3D_to_stream2D.c,
   image_crop2D.c, image_merge3D.c) to enable GCC inline
   architecture-specific memory move instructions.

2. Float precision fixes
   Replace double-precision math on float data with f-suffixed
   counterparts (cosf, sinf, sqrtf, atan2f) in complex conversion
   routines (image_mk_complex_from_amph.c, image_mk_amph_from_complex.c).
   Fix accidental powf() use for double data in imfunctions.c.

3. malloc/free elimination in compute loop
   Refactor image_slicenormalize.c to hoist working-array allocations
   out of the per-frame fpsexec() loop, replacing per-call malloc/free
   with persistent realloc()-managed buffers.

4. Array pointer hoisting + MILK_RESTRICT/MILK_ASSUME_ALIGNED
   In complex image conversion routines (image_mk_complex_from_amph.c,
   image_mk_amph_from_complex.c, image_mk_complex_from_reim.c,
   image_mk_reim_from_complex.c), hoist inner-loop array pointers to
   local MILK_RESTRICT + MILK_ASSUME_ALIGNED variables, enabling
   GCC to emit aligned AVX loads (vmovaps) and full SIMD vectorization.

5. MILK_ASSUME_ALIGNED on accumulation-loop restrict pointers
   Add MILK_ASSUME_ALIGNED to all typed restrict ptr declarations in
   arith_image_total_IMGID, arith_image_sumsquare_IMGID,
   arith_image_min_IMGID and arith_image_max_IMGID (image_total.c,
   image_stats.c).

6. Pointer hoisting in stream-diff inner loops
   Replace per-iteration struct dereferences in stream_diff.c and
   stream_halfimdiff.c with pre-hoisted MILK_RESTRICT/MILK_ASSUME_ALIGNED
   local pointer variables for all typed switch-case branches.

7. MILK_HOT annotations
   Mark arith_image_total_IMGID, arith_image_sumsquare_IMGID,
   arith_image_min_IMGID, arith_image_max_IMGID,
   COREMOD_MEMORY_streamDiff and COREMOD_MEMORY_stream_halfimDiff
   with MILK_HOT to guide linker co-location and aggressive inlining.

+++ src/coremods/COREMOD_arith/imfunctions.c
@@ -2148,7 +2148,7 @@ errno_t arith_image_cstpow_optimized_IMGID(IMGID *imgin, double f1, IMGID *imgou
             for(uint64_t i=0; i<nelement; i++) po[i] = p1[i] * p1[i];
         } else {
             _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = powf(p1[i], (float)f1);
+            for(uint64_t i=0; i<nelement; i++) po[i] = pow(p1[i], f1);
         }
     }
     else
